Fix block reg on server side (hotfix)
