+++ src/main/java/com/minecolonies/coremod/proxy/CommonProxy.java
@@ -2,6 +2,7 @@ package com.minecolonies.coremod.proxy;
 
 import com.minecolonies.api.util.constant.Constants;
 import com.minecolonies.coremod.MineColonies;
+import com.minecolonies.coremod.blocks.ModBlocks;
 import com.minecolonies.coremod.client.gui.WindowBuildTool;
 import com.minecolonies.coremod.colony.CitizenDataView;
 import com.minecolonies.coremod.entity.EntityCitizen;
@@ -13,16 +14,23 @@ import com.minecolonies.coremod.entity.ai.mobs.util.BarbarianSpawnUtils;
 import com.minecolonies.coremod.event.EventHandler;
 import com.minecolonies.coremod.event.FMLEventHandler;
 import com.minecolonies.coremod.inventory.GuiHandler;
+import com.minecolonies.coremod.items.ModItems;
 import com.minecolonies.coremod.sounds.ModSoundEvents;
 import com.minecolonies.coremod.tileentities.*;
+import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.storage.loot.LootTableList;
+import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.network.NetworkRegistry;
 import net.minecraftforge.fml.common.registry.EntityRegistry;
 import net.minecraftforge.fml.common.registry.GameRegistry;
+import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.util.HashMap;
@@ -31,6 +39,7 @@ import java.util.Map;
 /**
  * CommonProxy of the minecolonies mod (Server and Client).
  */
+@Mod.EventBusSubscriber
 public class CommonProxy implements IProxy
 {
     /**
@@ -67,6 +76,31 @@ public class CommonProxy implements IProxy
     public static NBTTagCompound getEntityData(final String name)
     {
         return playerPropertiesData.remove(name);
+    }
+
+    /**
+     * Called when registering blocks,
+     * we have to register all our modblocks here.
+     *
+     * @param event the registery event for blocks.
+     */
+    @SubscribeEvent
+    public static void registerBlocks(@NotNull final RegistryEvent.Register<Block> event)
+    {
+        ModBlocks.init(event.getRegistry());
+    }
+
+    /**
+     * Called when registering items,
+     * we have to register all our mod items here.
+     *
+     * @param event the registery event for items.
+     */
+    @SubscribeEvent
+    public static void registerItems(@NotNull final RegistryEvent.Register<Item> event)
+    {
+        ModItems.init(event.getRegistry());
+        ModBlocks.registerItemBlock(event.getRegistry());
     }
 
     @Override
